move PartiallyOrderedSet from gt-xml to gt-metadata (needed for Java 9 compatibility)
